requirements.txt
pandas
plotly
numpy
docxtpl==0.16.7
python-dotenv
requests==2.31.0